api/backend: change backend_mask_image.origin from ivec2 to vec2
A bit more precision for mask images. An attempt to reduce gaps when
window/shadow is scaled.

Changelog: Internal: Backend API 2.0 breaking change,
backend_mask_image.origin is now a vec2.

+++ include/picom/backend.h
@@ -92,7 +92,7 @@ struct backend_mask_image {
 	/// rounded.
 	double corner_radius;
 	/// Origin of the mask image, in the source image's coordinate.
-	ivec2 origin;
+	vec2 origin;
 	/// Whether the mask image should be inverted.
 	bool inverted;
 };
